requirements.txt
Django==1.8.3
Markdown==2.6.2
argparse==1.2.1
django-filter==0.11.0
djangorestframework==3.2.2
psycopg2==2.6.1
wsgiref==0.1.2
